⭐️ generate cask
